+++ src/app/api/v1/templates.py
@@ -5,21 +5,25 @@ from ...core.auth.auth import get_current_user
 from ...core.db.database import async_get_db
 from ...crud.crud_host_templates import (
     create_host_template,
+    delete_host_template,
     get_host_template,
     get_host_template_headers,
 )
 from ...crud.crud_range_templates import (
     create_range_template,
+    delete_range_template,
     get_range_template,
     get_range_template_headers,
 )
 from ...crud.crud_subnet_templates import (
     create_subnet_template,
+    delete_subnet_template,
     get_subnet_template,
     get_subnet_template_headers,
 )
 from ...crud.crud_vpc_templates import (
     create_vpc_template,
+    delete_vpc_template,
     get_vpc_template,
     get_vpc_template_headers,
 )
@@ -146,6 +150,48 @@ async def upload_range_template_endpoint(
     return TemplateRangeID.model_validate(created_range, from_attributes=True)
 
 
+@router.delete("/ranges/{range_id}")
+async def delete_range_template_endpoint(
+    range_id: str, db: AsyncSession = Depends(async_get_db)  # noqa: B008
+) -> bool:
+    """Delete a range template.
+
+    Args:
+    ----
+        range_id (str): Id of the range template.
+        db (AsyncSession): Async database connection.
+
+    Returns:
+    -------
+        bool: True if successfully deleted. False otherwise.
+
+    """
+    # Invalid UUID4 ID
+    if not is_valid_uuid4(range_id):
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="ID provided is not a valid UUID4.",
+        )
+
+    range_template = await get_range_template(db, TemplateRangeID(id=range_id))
+
+    # Does not exist
+    if not range_template:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Range template with id: {range_id} not found!",
+        )
+
+    # Not standalone template
+    if not range_template.is_standalone():
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail="Cannot delete range template because it is not a standalone template.",
+        )
+
+    return await delete_range_template(db, range_template)
+
+
 @router.get("/vpcs")
 async def get_vpc_template_headers_endpoint(
     standalone_only: bool = True,
@@ -243,6 +289,48 @@ async def upload_vpc_template_endpoint(
     # Create the template with the current user as the owner
     created_vpc = await create_vpc_template(db, vpc_template, owner_id=current_user.id)
     return TemplateVPCID.model_validate(created_vpc, from_attributes=True)
+
+
+@router.delete("/vpcs/{vpc_id}")
+async def delete_vpc_template_endpoint(
+    vpc_id: str, db: AsyncSession = Depends(async_get_db)  # noqa: B008
+) -> bool:
+    """Delete a VPC template.
+
+    Args:
+    ----
+        vpc_id (str): Id of the VPC template.
+        db (AsyncSession): Async database connection.
+
+    Returns:
+    -------
+        bool: True if successfully deleted. False otherwise.
+
+    """
+    # Invalid UUID4 ID
+    if not is_valid_uuid4(vpc_id):
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="ID provided is not a valid UUID4.",
+        )
+
+    vpc_template = await get_vpc_template(db, TemplateVPCID(id=vpc_id))
+
+    # Does not exist
+    if not vpc_template:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"VPC template with id: {vpc_id} not found!",
+        )
+
+    # Not standalone template
+    if not vpc_template.is_standalone():
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"Cannot delete VPC template because it is not a standalone template. Connected to range: {vpc_template.range_id}",
+        )
+
+    return await delete_vpc_template(db, vpc_template)
 
 
 @router.get("/subnets")
@@ -346,6 +434,48 @@ async def upload_subnet_template_endpoint(
     return TemplateSubnetID.model_validate(created_subnet, from_attributes=True)
 
 
+@router.delete("/subnets/{subnet_id}")
+async def delete_subnet_template_endpoint(
+    subnet_id: str, db: AsyncSession = Depends(async_get_db)  # noqa: B008
+) -> bool:
+    """Delete a subnet template.
+
+    Args:
+    ----
+        subnet_id (str): Id of the subnet template.
+        db (AsyncSession): Async database connection.
+
+    Returns:
+    -------
+        bool: True if successfully deleted. False otherwise.
+
+    """
+    # Invalid UUID4 ID
+    if not is_valid_uuid4(subnet_id):
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="ID provided is not a valid UUID4.",
+        )
+
+    subnet_template = await get_subnet_template(db, TemplateSubnetID(id=subnet_id))
+
+    # Does not exist
+    if not subnet_template:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Subnet template with id: {subnet_id} not found!",
+        )
+
+    # Not standalone template
+    if not subnet_template.is_standalone():
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"Cannot delete subnet template because it is not a standalone template. Connected to VPC: {subnet_template.vpc_id}",
+        )
+
+    return await delete_subnet_template(db, subnet_template)
+
+
 @router.get("/hosts")
 async def get_host_template_headers_endpoint(
     standalone_only: bool = True,
@@ -445,3 +575,46 @@ async def upload_host_template_endpoint(
     )
 
     return TemplateHostID.model_validate(created_host, from_attributes=True)
+
+
+
+@router.delete("/hosts/{host_id}")
+async def delete_host_template_endpoint(
+    host_id: str, db: AsyncSession = Depends(async_get_db)  # noqa: B008
+) -> bool:
+    """Delete a host template.
+
+    Args:
+    ----
+        host_id (str): Id of the host.
+        db (AsyncSession): Async database connection.
+
+    Returns:
+    -------
+        bool: True if successfully deleted. False otherwise.
+
+    """
+    # Invalid UUID4 ID
+    if not is_valid_uuid4(host_id):
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="ID provided is not a valid UUID4.",
+        )
+
+    host_template = await get_host_template(db, TemplateHostID(id=host_id))
+
+    # Does not exist
+    if not host_template:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Host with id: {host_id} not found!",
+        )
+
+    # Not standalone template
+    if not host_template.is_standalone():
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"Cannot delete host template because it is not a standalone template. Connected to subnet: {host_template.subnet_id}",
+        )
+
+    return await delete_host_template(db, host_template)