+++ src/app/models/template_base_model.py
@@ -5,7 +5,7 @@ from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, MappedAsDataclass, mapped_column
 
 
-class OpenLabsTemplateMixin(MappedAsDataclass):
+class TemplateModelMixin(MappedAsDataclass):
     """Mixin to provide a UUID for each template-based model."""
 
     id: Mapped[uuid.UUID] = mapped_column(