+++ tests/validators/test_id.py
@@ -1,4 +1,5 @@
 import uuid
+from random import randint
 
 from src.app.validators.id import is_valid_uuid4
 
@@ -15,3 +16,9 @@ def test_invalid_uuid4() -> None:
     for _ in range(5):
         invalid_uuid = str(uuid.uuid4())[:-1]
         assert not is_valid_uuid4(invalid_uuid)
+
+
+def test_wrong_uuid_version_uuid4() -> None:
+    """Test that we get false for the wrong UUID version."""
+    uuid3_str = str(uuid.uuid1(randint(0, 100), randint(0, 100)))  # noqa: S311
+    assert not is_valid_uuid4(uuid3_str)