+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# v0.6.2 (Thu Feb 27 2025)
+
+#### 🐛 Bug Fix
+
+- Add template delete endpoints [#55](https://github.com/OpenLabsX/API/pull/55) ([@alexchristy](https://github.com/alexchristy))
+
+#### Authors: 1
+
+- Alex Christy ([@alexchristy](https://github.com/alexchristy))
+
+---
+
 # v0.6.1 (Sun Feb 23 2025)
 
 #### 🐛 Bug Fix